+++ ferry_store/lib/src/store.dart
@@ -1,22 +1,17 @@
+import 'package:rxdart/rxdart.dart';
+
 abstract class Store {
-  Stream<Map<String, Map<String, dynamic>>> watch();
+  ValueStream<Map<String, Map<String, dynamic>>> valueStream;
 
-  Map<String, dynamic> get(
-    String dataId,
-  );
+  Map<String, dynamic> get(String dataId);
 
-  void put(
-    String dataId,
-    Map<String, dynamic> value,
-  );
+  void put(String dataId, Map<String, dynamic> value);
 
-  void putAll(
-    Map<String, Map<String, dynamic>> data,
-  );
+  void putAll(Map<String, Map<String, dynamic>> data);
 
-  void delete(
-    String dataId,
-  );
+  void delete(String dataId);
 
-  Map<String, Map<String, dynamic>> toMap();
+  void clear();
+
+  void dispose();
 }