+++ ferry_hive_store/CHANGELOG.md
@@ -12,11 +12,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - add tests
+- add clear method
+- add dispose method
 
 ### Changed
 
 - [breaking] use typed box to avoid necessity to recast
 - base watch stream on underlying hive watch
+
+### Removed
+
+- remove `toMap()` method
 
 ## [0.1.0] - 2020-05-10